WIP dependent file checks
